requirements.txt
alpaca-py>=0.39.0
apscheduler>=3.10.4
fastapi>=0.115.0
pandas>=2.2.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.0
streamlit>=1.37.0
uvicorn[standard]>=0.30.0
pytest>=8.3.0
